requirements.txt
pybit>=5.6.0
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
anthropic>=0.30.0
google-genai>=1.0.0
openai>=1.30.0
